// generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Address = Node & {
	__typename?: 'Address';
	city?: Maybe<Scalars['String']['output']>;
	company?: Maybe<Scalars['String']['output']>;
	country: Country;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	defaultBillingAddress?: Maybe<Scalars['Boolean']['output']>;
	defaultShippingAddress?: Maybe<Scalars['Boolean']['output']>;
	fullName?: Maybe<Scalars['String']['output']>;
	id: Scalars['ID']['output'];
	phoneNumber?: Maybe<Scalars['String']['output']>;
	postalCode?: Maybe<Scalars['String']['output']>;
	province?: Maybe<Scalars['String']['output']>;
	streetLine1: Scalars['String']['output'];
	streetLine2?: Maybe<Scalars['String']['output']>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Asset = Node & {
	__typename?: 'Asset';
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	fileSize: Scalars['Int']['output'];
	focalPoint?: Maybe<Coordinate>;
	height: Scalars['Int']['output'];
	id: Scalars['ID']['output'];
	mimeType: Scalars['String']['output'];
	name: Scalars['String']['output'];
	preview: Scalars['String']['output'];
	source: Scalars['String']['output'];
	tags: Array<Tag>;
	type: AssetType;
	updatedAt: Scalars['DateTime']['output'];
	width: Scalars['Int']['output'];
};
// ...
export type Country = Node &
	Region & {
		__typename?: 'Country';
		code: Scalars['String']['output'];
		createdAt: Scalars['DateTime']['output'];
		customFields?: Maybe<Scalars['JSON']['output']>;
		enabled: Scalars['Boolean']['output'];
		id: Scalars['ID']['output'];
		languageCode: LanguageCode;
		name: Scalars['String']['output'];
		parent?: Maybe<Region>;
		parentId?: Maybe<Scalars['ID']['output']>;
		translations: Array<RegionTranslation>;
		type: Scalars['String']['output'];
		updatedAt: Scalars['DateTime']['output'];
	};
// ...
export type CreateAddressInput = {
	city?: InputMaybe<Scalars['String']['input']>;
	company?: InputMaybe<Scalars['String']['input']>;
	countryCode: Scalars['String']['input'];
	customFields?: InputMaybe<Scalars['JSON']['input']>;
	defaultBillingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	defaultShippingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	fullName?: InputMaybe<Scalars['String']['input']>;
	phoneNumber?: InputMaybe<Scalars['String']['input']>;
	postalCode?: InputMaybe<Scalars['String']['input']>;
	province?: InputMaybe<Scalars['String']['input']>;
	streetLine1: Scalars['String']['input'];
	streetLine2?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type CurrencyCode = (typeof CurrencyCode)[keyof typeof CurrencyCode];
// ...
export type Customer = Node & {
	__typename?: 'Customer';
	addresses?: Maybe<Array<Address>>;
	createdAt: Scalars['DateTime']['output'];
	customFields?: Maybe<Scalars['JSON']['output']>;
	emailAddress: Scalars['String']['output'];
	firstName: Scalars['String']['output'];
	id: Scalars['ID']['output'];
	lastName: Scalars['String']['output'];
	orders: OrderList;
	phoneNumber?: Maybe<Scalars['String']['output']>;
	title?: Maybe<Scalars['String']['output']>;
	updatedAt: Scalars['DateTime']['output'];
	user?: Maybe<User>;
};
// ...
export type ErrorCode = (typeof ErrorCode)[keyof typeof ErrorCode];
// ...
export type InvalidCredentialsError = ErrorResult & {
	__typename?: 'InvalidCredentialsError';
	authenticationError: Scalars['String']['output'];
	errorCode: ErrorCode;
	message: Scalars['String']['output'];
};
// ...
export type Node = {
	id: Scalars['ID']['output'];
};
// ...
export type Order = Node & {
	__typename?: 'Order';
	/** An order is active as long as the payment process has not been completed */
	active: Scalars['Boolean']['output'];
	billingAddress?: Maybe<OrderAddress>;
	/** A unique code for the Order */
	code: Scalars['String']['output'];
	/** An array of all coupon codes applied to the Order */
	couponCodes: Array<Scalars['String']['output']>;
	createdAt: Scalars['DateTime']['output'];
	currencyCode: CurrencyCode;
	customFields?: Maybe<Scalars['JSON']['output']>;
	customer?: Maybe<Customer>;
	discounts: Array<Discount>;
	fulfillments?: Maybe<Array<Fulfillment>>;
	history: HistoryEntryList;
	id: Scalars['ID']['output'];
	lines: Array<OrderLine>;
	/**
	 * The date & time that the Order was placed, i.e. the Customer
	 * completed the checkout and the Order is no longer "active"
	 */
	orderPlacedAt?: Maybe<Scalars['DateTime']['output']>;
	payments?: Maybe<Array<Payment>>;
	/** Promotions applied to the order. Only gets populated after the payment process has completed. */
	promotions: Array<Promotion>;
	shipping: Scalars['Money']['output'];
	shippingAddress?: Maybe<OrderAddress>;
	shippingLines: Array<ShippingLine>;
	shippingWithTax: Scalars['Money']['output'];
	state: Scalars['String']['output'];
	/**
	 * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
	 * discounts which have been prorated (proportionally distributed) amongst the items of each OrderLine.
	 * To get a total of all OrderLines which does not account for prorated discounts, use the
	 * sum of `OrderLine.discountedLinePrice` values.
	 */
	subTotal: Scalars['Money']['output'];
	/** Same as subTotal, but inclusive of tax */
	subTotalWithTax: Scalars['Money']['output'];
	/**
	 * Surcharges are arbitrary modifications to the Order total which are neither
	 * ProductVariants nor discounts resulting from applied Promotions. For example,
	 * one-off discounts based on customer interaction, or surcharges based on payment
	 * methods.
	 */
	surcharges: Array<Surcharge>;
	/** A summary of the taxes being applied to this Order */
	taxSummary: Array<OrderTaxSummary>;
	/** Equal to subTotal plus shipping */
	total: Scalars['Money']['output'];
	totalQuantity: Scalars['Int']['output'];
	/** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
	totalWithTax: Scalars['Money']['output'];
	type: OrderType;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type OrderList = PaginatedList & {
	__typename?: 'OrderList';
	items: Array<Order>;
	totalItems: Scalars['Int']['output'];
};
// ...
export type ProductVariant = Node & {
	__typename?: 'ProductVariant';
	assets: Array<Asset>;
	createdAt: Scalars['DateTime']['output'];
	currencyCode: CurrencyCode;
	customFields?: Maybe<Scalars['JSON']['output']>;
	facetValues: Array<FacetValue>;
	featuredAsset?: Maybe<Asset>;
	id: Scalars['ID']['output'];
	languageCode: LanguageCode;
	name: Scalars['String']['output'];
	options: Array<ProductOption>;
	price: Scalars['Money']['output'];
	priceWithTax: Scalars['Money']['output'];
	product: Product;
	productId: Scalars['ID']['output'];
	sku: Scalars['String']['output'];
	stockLevel: Scalars['String']['output'];
	taxCategory: TaxCategory;
	taxRateApplied: TaxRate;
	translations: Array<ProductVariantTranslation>;
	updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Query = {
	__typename?: 'Query';
	/** The active Channel */
	activeChannel: Channel;
	/** The active Customer */
	activeCustomer?: Maybe<Customer>;
	/**
	 * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
	 * state of `PaymentAuthorized` or `PaymentSettled`, then that Order is no longer considered "active" and this
	 * query will once again return `null`.
	 */
	activeOrder?: Maybe<Order>;
	/** An array of supported Countries */
	availableCountries: Array<Country>;
	/** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is specified, an error will result. */
	collection?: Maybe<Collection>;
	/** A list of Collections available to the shop */
	collections: CollectionList;
	/** Returns a list of payment methods and their eligibility based on the current active Order */
	eligiblePaymentMethods: Array<PaymentMethodQuote>;
	/** Returns a list of eligible shipping methods based on the current active Order */
	eligibleShippingMethods: Array<ShippingMethodQuote>;
	/** Returns a Facet by its id */
	facet?: Maybe<Facet>;
	/** A list of Facets available to the shop */
	facets: FacetList;
	generateBraintreeClientToken?: Maybe<Scalars['String']['output']>;
	/** Returns information about the current authenticated User */
	me?: Maybe<CurrentUser>;
	/** Returns the possible next states that the activeOrder can transition to */
	nextOrderStates: Array<Scalars['String']['output']>;
	/**
	 * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
	 * currently-authenticated User may be queried.
	 */
	order?: Maybe<Order>;
	/**
	 * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
	 * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
	 * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
	 * general anonymous access to Order data.
	 */
	orderByCode?: Maybe<Order>;
	/** Get a Product either by id or slug. If neither 'id' nor 'slug' is specified, an error will result. */
	product?: Maybe<Product>;
	/** Get a list of Products */
	products: ProductList;
	/** Search Products based on the criteria set by the `SearchInput` */
	search: SearchResponse;
};
// ...
export type UpdateAddressInput = {
	city?: InputMaybe<Scalars['String']['input']>;
	company?: InputMaybe<Scalars['String']['input']>;
	countryCode?: InputMaybe<Scalars['String']['input']>;
	customFields?: InputMaybe<Scalars['JSON']['input']>;
	defaultBillingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	defaultShippingAddress?: InputMaybe<Scalars['Boolean']['input']>;
	fullName?: InputMaybe<Scalars['String']['input']>;
	id: Scalars['ID']['input'];
	phoneNumber?: InputMaybe<Scalars['String']['input']>;
	postalCode?: InputMaybe<Scalars['String']['input']>;
	province?: InputMaybe<Scalars['String']['input']>;
	streetLine1?: InputMaybe<Scalars['String']['input']>;
	streetLine2?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type ActiveCustomerAddressesQuery = {
	__typename?: 'Query';
	activeCustomer?: {
		__typename?: 'Customer';
		id: string;
		addresses?: Array<{
			__typename?: 'Address';
			id: string;
			fullName?: string | null;
			company?: string | null;
			streetLine1: string;
			streetLine2?: string | null;
			city?: string | null;
			province?: string | null;
			postalCode?: string | null;
			phoneNumber?: string | null;
			defaultShippingAddress?: boolean | null;
			defaultBillingAddress?: boolean | null;
			country: { __typename?: 'Country'; code: string };
		}> | null;
	} | null;
};

export type ActiveCustomerQueryVariables = Exact<{ [key: string]: never }>;

export type ActiveCustomerQuery = {
	__typename?: 'Query';
	activeCustomer?: {
		__typename?: 'Customer';
		id: string;
		title?: string | null;
		firstName: string;
		lastName: string;
		emailAddress: string;
		phoneNumber?: string | null;
	} | null;
};
// ...
export type ActiveCustomerOrdersQueryVariables = Exact<{
	options?: InputMaybe<OrderListOptions>;
}>;

export type ActiveCustomerOrdersQuery = {
	__typename?: 'Query';
	activeCustomer?: {
		__typename?: 'Customer';
		id: string;
		orders: {
			__typename?: 'OrderList';
			totalItems: number;
			items: Array<{
				__typename?: 'Order';
				id: string;
				code: string;
				state: string;
				totalWithTax: any;
				currencyCode: CurrencyCode;
				lines: Array<{
					__typename?: 'OrderLine';
					featuredAsset?: { __typename?: 'Asset'; preview: string } | null;
					productVariant: { __typename?: 'ProductVariant'; name: string };
				}>;
			}>;
		};
	} | null;
};

export type UpdateCustomerPasswordMutationMutationVariables = Exact<{
	currentPassword: Scalars['String']['input'];
	newPassword: Scalars['String']['input'];
}>;

export type UpdateCustomerPasswordMutationMutation = {
	__typename?: 'Mutation';
	updateCustomerPassword:
		| { __typename: 'InvalidCredentialsError'; errorCode: ErrorCode; message: string }
		| { __typename: 'NativeAuthStrategyError'; errorCode: ErrorCode; message: string }
		| { __typename: 'PasswordValidationError'; errorCode: ErrorCode; message: string }
		| { __typename: 'Success'; success: boolean };
};